attach reused utilities to ctfd
